refactor: kakao 컴포넌트를 suspense처리
